+++ Project17-C-Map/InteractiveClusteringMap/Clustering/ClusteringServicing.swift
@@ -13,8 +13,5 @@ protocol ClusteringServicing {
                  boundingBox: BoundingBox,
                  zoomLevel: Double,
                  completionHandler: @escaping (([Cluster]) -> Void))
-    func cancel()
-    func add(coordinate: Coordinate)
-    func delete(coordinate: Coordinate)
     
 }
